Moved addNewAccount function to a new requests utility file
